Support for development builds, macOS 10.11 local networking support

+++ buildAndSign.sh
@@ -24,16 +24,21 @@ mkdir build/libs/tmp
 pushd build/libs/tmp
 
 #Sign native JAR libraries
-for f in ../*.jar;
-do
-	echo "Re-signing $(basename "$f")"
+if [ -z "$SIGNATURE" ]
+then
+	echo "Skipping re-signing dependencies"
+else
+	for f in ../*.jar;
+	do
+		echo "Re-signing $(basename "$f")"
 
-	jar xf "$f" #Unpack
-	rm "$f" #Delete original JAR
-	find -E . -regex ".*\.(dylib|jnilib)" -print0 | xargs codesign --force --verbose --sign "$SIGNATURE" #Codesign dynamic libraries
-	jar cmf META-INF/MANIFEST.MF "$f" ./* #Repack JAR
-	rm -r ./* #Empty directory
-done
+		jar xf "$f" #Unpack
+		rm "$f" #Delete original JAR
+		find -E . -regex ".*\.(dylib|jnilib)" -print0 | xargs codesign --force --verbose --sign "$SIGNATURE" #Codesign dynamic libraries
+		jar cmf META-INF/MANIFEST.MF "$f" ./* #Repack JAR
+		rm -r ./* #Empty directory
+	done
+fi
 
 #Clean up tmp directory
 popd
@@ -62,51 +67,67 @@ $JAVA_HOME/bin/jpackage \
 #Update app plist
 echo "Fixing plist"
 plutil -insert LSUIElement -string True "$APP_FILE/Contents/Info.plist" #Hide dock icon
-plutil -insert NSAppTransportSecurity -xml "<dict><key>NSAllowsLocalNetworking</key><true/></dict>" "$APP_FILE/Contents/Info.plist" #Enable local networking (for AirMessage Connect sign-in)
+plutil -insert NSAppTransportSecurity -xml "<dict><key>NSAllowsLocalNetworking</key><true/><key>NSAllowsArbitraryLoads</key><true/></dict>" "$APP_FILE/Contents/Info.plist" #Enable local networking (for AirMessage Connect sign-in)
 
 #Sign app
-echo "Signing app"
-codesign --force --options runtime --entitlements "macos.entitlements" --sign "$SIGNATURE" "$APP_FILE/Contents/runtime/Contents/MacOS/libjli.dylib"
-codesign --force --options runtime --entitlements "macos.entitlements" --sign "$SIGNATURE" "$APP_FILE/Contents/MacOS/AirMessage"
-codesign --force --options runtime --entitlements "macos.entitlements" --sign "$SIGNATURE" "$APP_FILE"
+if [ -z "$SIGNATURE" ]
+then
+	echo "Skipping signing app"
+else
+	echo "Signing app"
+	codesign --force --options runtime --entitlements "macos.entitlements" --sign "$SIGNATURE" "$APP_FILE/Contents/runtime/Contents/MacOS/libjli.dylib"
+	codesign --force --options runtime --entitlements "macos.entitlements" --sign "$SIGNATURE" "$APP_FILE/Contents/MacOS/AirMessage"
+	codesign --force --options runtime --entitlements "macos.entitlements" --sign "$SIGNATURE" "$APP_FILE"
+fi
 
 #Package app to ZIP
-echo "Compressing app for notarization"
+if [ -z "$NOTARIZATION_PASSKEY" ]
+then
+	echo "Compressing app for development"
+else
+	echo "Compressing app for notarization"
+fi
 ditto -c -k --keepParent "$APP_FILE" "$PACKAGE_FILE"
 
-#Notarize app
-echo "Uploading app to Apple notarization service"
-REQUEST_UUID=$(xcrun altool --notarize-app \
-	--primary-bundle-id "me.tagavari.airmessageserver" \
-	--username $NOTARIZATION_USERNAME \
-	--password "@keychain:$NOTARIZATION_PASSKEY" \
-	--file "$PACKAGE_FILE" \
-	| grep RequestUUID | awk '{print $3}')
-rm "$PACKAGE_FILE"
+if [ -z "$NOTARIZATION_PASSKEY" ]
+then
+	echo "Skipping notarization"
+	echo "Successfully built AirMessage Server v$VERSION for development"
+else
+	#Notarize app
+	echo "Uploading app to Apple notarization service"
+	REQUEST_UUID=$(xcrun altool --notarize-app \
+		--primary-bundle-id "me.tagavari.airmessageserver" \
+		--username $NOTARIZATION_USERNAME \
+		--password "@keychain:$NOTARIZATION_PASSKEY" \
+		--file "$PACKAGE_FILE" \
+		| grep RequestUUID | awk '{print $3}')
+	rm "$PACKAGE_FILE"
 
-#Wait for notarization to finish
-echo "Waiting for completion of notarization request $REQUEST_UUID"
-while true; do
-	NOTARIZATION_STATUS=$(xcrun altool --notarization-info "$REQUEST_UUID" --username "$NOTARIZATION_USERNAME" --password "@keychain:$NOTARIZATION_PASSKEY")
-	if echo "$NOTARIZATION_STATUS" | grep -q "Status: in progress"; then sleep 20
-	elif echo "$NOTARIZATION_STATUS" | grep -q "Status: success"; then break
-	else
-		>&2 echo "$NOTARIZATION_STATUS"
-		exit
-	fi
-done
+	#Wait for notarization to finish
+	echo "Waiting for completion of notarization request $REQUEST_UUID"
+	while true; do
+		NOTARIZATION_STATUS=$(xcrun altool --notarization-info "$REQUEST_UUID" --username "$NOTARIZATION_USERNAME" --password "@keychain:$NOTARIZATION_PASSKEY")
+		if echo "$NOTARIZATION_STATUS" | grep -q "Status: in progress"; then sleep 20
+		elif echo "$NOTARIZATION_STATUS" | grep -q "Status: success"; then break
+		else
+			>&2 echo "$NOTARIZATION_STATUS"
+			exit
+		fi
+	done
 
-#Staple ticket
-echo "Stapling ticket"
-xcrun stapler staple "$APP_FILE"
+	#Staple ticket
+	echo "Stapling ticket"
+	xcrun stapler staple "$APP_FILE"
 
-#Check for signatures
-echo "Verifying files"
-spctl --assess "$APP_FILE"
-codesign --verify "$APP_FILE"
+	#Check for signatures
+	echo "Verifying files"
+	spctl --assess "$APP_FILE"
+	codesign --verify "$APP_FILE"
 
-#Re-compress app
-echo "Compressing final app to $PACKAGE_FILE"
-ditto -c -k --keepParent "$APP_FILE" "$PACKAGE_FILE"
+	#Re-compress app
+	echo "Compressing final app to $PACKAGE_FILE"
+	ditto -c -k --keepParent "$APP_FILE" "$PACKAGE_FILE"
 
-echo "Successfully built AirMessage Server v$VERSION for distribution"
+	echo "Successfully built AirMessage Server v$VERSION for distribution"
+fi
